README.md
A Python-based network port scanner for MacOS Unix terminal is capable of scanning a given website IP for open ports. The tool can help to identify security vulnerabilities and ensure network safety. It provides a comprehensive list of open ports on the target system, helping to Identify potential attack vectors. The network port scanner is easy to use, and the output is generated in a concise and readable format. This makes it a valuable tool for both network administrators and cybersecurity professionals to secure their networks.